process video segments individually, autoqueue and manual processing
per segment processing. 
the workflow generates a single video segment per run, auto queues the necessary runs, allows to manually regenerate specific segments, instructions sent to the LLM are simpler (only one scene is sent to the LLM but it does mean that more tokens are consumed as the instructions boilerplate is repeated for each scene) so a simple local LLM is able to handle it. the last audio segment gets padded with silence if necessary to fill the 4 seconds minimum per segment
Its a big mess right now in need of some clean up but works to prove the idea